+++ RELEASE_NOTES.md
@@ -13,6 +13,8 @@
     * The constructor parameter `channel_options` was renamed to `channels_defaults` to match the name used in `BaseApiClient`.
     * The constructor now accepts a `connect` parameter, which is `True` by default. If set to `False`, the client will not connect to the server upon instantiation. You can connect later by calling the `connect()` method.
 
+* The `frequenz-client-base` dependency was bumped to v0.8.0.
+
 ## New Features
 
 - The client now inherits from `frequenz.client.base.BaseApiClient`, so it provides a few new features, like `disconnect()`ing or using it as a context manager. Please refer to the [`BaseApiClient` documentation](https://frequenz-floss.github.io/frequenz-client-base-python/latest/reference/frequenz/client/base/client/#frequenz.client.base.client.BaseApiClient) for more information on these features.